Split DataFileInterface tests into separate file
